+++ test/dart_date_test.dart
@@ -2,10 +2,27 @@ import 'package:dart_date/dart_date.dart';
 import 'package:test/test.dart';
 
 void main() {
-  group('First', () {
-    test('Second', () {
-      Date d = Date.today;
-      expect(d, d.add(Duration(days: 0)));
+  group('Getters', () {
+    test('timestamp', (){
+      expect(Date.unix(0).timestamp, 0);
+      expect(Date.parse("1996-03-29T11:11:11.011Z").timestamp, 828097871011);
+    });
+
+    test('isFirstDayOfMonth', () {
+      expect(Date(2011, 2, 1, 11).isFirstDayOfMonth, true);
+      expect(Date.parse('November 01 2018, 9:14:29 PM', pattern: 'MMMM dd y, h:mm:ss a').isFirstDayOfMonth, true);
+      expect(Date.parse('November 30 2011, 0:14:29 PM', pattern: 'MMMM dd y, h:mm:ss a').isFirstDayOfMonth, false);
+    });
+
+    test('isLastDayOfMonth', () {
+      expect(Date(2011, 2, 1, 11).isLastDayOfMonth, false);
+      expect(Date.parse('November 01 2018, 9:14:29 PM', pattern: 'MMMM dd y, h:mm:ss a').isLastDayOfMonth, false);
+      expect(Date.parse('November 30 2011, 0:14:29 PM', pattern: 'MMMM dd y, h:mm:ss a').isLastDayOfMonth, true);
+    });
+    
+    test('isLeapYear', () {
+      expect(Date(2011, 2, 1, 11).isLeapYear, false);
+      expect(Date.parse('September 12 2012', pattern: 'MMMM dd y').isLeapYear, true);
     });
   });
 }