+++ CHANGELOG.md
@@ -1,4 +1,9 @@
 # Changelog
 
+## 0.0.9
+- Changes `Date.parse` to accept or not a format
+- Added `date.isFirstDayOfMonth` getter
+- Added `date.isLastDayOfMonth` getter
+- Added `date.isLeapYear` getter
 ## 0.0.8
 - Adding documentation and timeago `date.timeago()`